Reworked the CSS for movies and vote pages to improve readability. Finished writing vote javascript to show all movie options. Added some extra error checking to vote.js.
