+++ movienightbot/webfiles/content/styles/movies.css
@@ -23,12 +23,12 @@
 
 .section.content .sorting .button {
 	padding: 15px;
-    background-color: #254770;
-    margin-left: 5px;
-    margin-right: 5px;
-    border-radius: 5px;
+	background-color: #254770;
+	margin-left: 5px;
+	margin-right: 5px;
+	border-radius: 5px;
 
-    cursor: pointer;
+	cursor: pointer;
 	-webkit-user-select: none; /* Safari */
 	-ms-user-select: none; /* IE 10 and IE 11 */
 	user-select: none; /* Standard syntax */
@@ -54,13 +54,13 @@
 }
 
 .section.content .letterNav {
-    display: flex;
-    float: right;
-    margin-top: 25px;
+	display: flex;
+	float: right;
+	margin-top: 25px;
 }
 
 .section.content .letterNav * {
-    cursor: pointer;
+	cursor: pointer;
 	-webkit-user-select: none; /* Safari */
 	-ms-user-select: none; /* IE 10 and IE 11 */
 	user-select: none; /* Standard syntax */
@@ -77,18 +77,18 @@
 }
 
 .section.content .controls {
-    display: flex;
-    margin-top: 10px;
+	display: flex;
+	margin-top: 10px;
 }
 
 .section.content .controls .button {
-    padding: 15px;
-    background-color: #254770;
-    margin-left: 5px;
-    margin-right: 5px;
-    border-radius: 5px;
+	padding: 15px;
+	background-color: #254770;
+	margin-left: 5px;
+	margin-right: 5px;
+	border-radius: 5px;
 
-    cursor: pointer;
+	cursor: pointer;
 	-webkit-user-select: none; /* Safari */
 	-ms-user-select: none; /* IE 10 and IE 11 */
 	user-select: none; /* Standard syntax */
@@ -96,10 +96,10 @@
 }
 
 .section.content .controls .label {
-    padding: 15px;
-    margin-left: 5px;
-    margin-right: 5px;
-    border-radius: 5px;
+	padding: 15px;
+	margin-left: 5px;
+	margin-right: 5px;
+	border-radius: 5px;
 
 	-webkit-user-select: none; /* Safari */
 	-ms-user-select: none; /* IE 10 and IE 11 */
@@ -113,10 +113,10 @@
 }
 
 .section.content .items .item {
-    display: flex;
-    justify-content: space-around;
-    margin-top: 10px;
-    margin-bottom: 10px;
+	display: flex;
+	justify-content: space-between;
+	margin-top: 10px;
+	margin-bottom: 10px;
 }
 
 /* .section.content .items .item #image {
@@ -124,29 +124,43 @@
 } */
 
 .section.content .items .item .coverImage {
-    width: 200px;
-    height: 200px;
-    object-fit: contain;
+	width: 200px;
+	height: 200px;
+	object-fit: contain;
+}
+
+.section.content .items .item #txtYear {
+	margin-top: 5px;
+	text-align: center;
+	margin-bottom: 15px;
 }
 
 .section.content .items .item .genres {
-	width: 20%;
+	width: 10%;
 }
 
 .section.content .items .item .imdbData {
-    width: 15%;
-    padding-left: 15px;
-    padding-right: 15px;
+	width: 20%;
+	padding-left: 15px;
+	padding-right: 15px;
+}
+
+.section.content .items .item #txtTitle {
+	display: -webkit-box;
+	-webkit-line-clamp: 4;
+	-webkit-box-orient: vertical;  
+	margin-top: 0px;
+	overflow: hidden;
 }
 
 .section.content .items .item .imdbData a {
-    text-decoration: none;
-    color: #D9E9FC;
+	text-decoration: none;
+	color: #D9E9FC;
 }
 
 .section.content .items .item .imdbData a:hover {
-    transition: 0.2s;
-    color: #5385C2;
+	transition: 0.2s;
+	color: #5385C2;
 }
 
 @media screen and (max-width: 1150px) {
@@ -162,6 +176,12 @@
 
 	.section.content .items .item .genres {
 		display: none;
+	}
+
+	.section.content .items .item .imdbData {
+		width: 30%;
+		padding-left: 15px;
+		padding-right: 15px;
 	}
 
 }
