+++ movienightbot/webfiles/content/styles/vote.css
@@ -1,0 +1,90 @@
+.section.content {
+	background-color: #263D59;
+	min-height: 75%;
+	color: #D9E9FC;
+	padding-left: 15%;
+	padding-right: 15%;
+	padding-top: 15px;
+	padding-bottom: 15px;
+	display: block;
+}
+
+.section.content .piechart {
+	display: flex;
+	justify-content: center;
+}
+
+.section.content .listHeader {
+	margin-top: 45px;
+	margin-bottom: 45px;
+	width: 100%;
+	text-align: center;
+}
+
+.section.content .movieList .item {
+	display: flex;
+	justify-content: space-between;
+	margin-top: 10px;
+	margin-bottom: 10px;
+}
+
+/* .section.content .items .item #image {
+	width: 20%;
+} */
+
+.section.content .movieList .item .coverImage {
+	width: 100px;
+	height: 100px;
+	object-fit: contain;
+}
+
+.section.content .movieList .item #txtYear {
+	margin-top: 5px;
+	text-align: center;
+}
+
+.section.content .movieList .item #txtTitle {
+	display: -webkit-box;
+	-webkit-line-clamp: 4;
+	-webkit-box-orient: vertical;  
+	margin-top: 0px;
+	overflow: hidden;
+}
+
+.section.content .movieList .item .genres {
+	width: 10%;
+}
+
+.section.content .movieList .item .imdbData {
+	width: 20%;
+	padding-left: 15px;
+	padding-right: 15px;
+}
+
+.section.content .movieList .item .imdbData a {
+	text-decoration: none;
+	color: #D9E9FC;
+}
+
+.section.content .movieList .item .imdbData a:hover {
+	transition: 0.2s;
+	color: #5385C2;
+}
+
+@media screen and (max-width: 1150px) {
+
+	.section.content .movieList .item .data2 {
+		display: none;
+	}
+
+	.section.content .movieList .item .genres {
+		display: none;
+	}
+
+	.section.content .movieList .item .imdbData {
+		width: 30%;
+		padding-left: 15px;
+		padding-right: 15px;
+	}
+
+}
