+++ movienightbot/webfiles/movies.html
@@ -19,6 +19,7 @@
 				<h1 class="title" id="TxtPageTitle">Movie Night Bot Suggestions</h1>
 			</a>
 			<div class="navigation">
+				<a class="link" id="BtnVote" onclick="javascript: MovieNightBot.OnClickBtnVote()"><h2>Show Vote</h2></a>
 				<a class="link" id="BtnSetting" onclick="javascript: MovieNightBot.OnClickBtnSetting()"><h2>Show Watched</h2></a>
 			</div>
 		</div>
